Fix whitespace between Purpose icons
Closes #41. The solution is purely CSS based, but requires precise definition of:
which breakpoint, at which size, and how many columns.

We're using a 'flexbox with wrap' and manually computing the width so that the
wrapping does not leave blank space and the outer box is shrunk to contents.

This is the only CSS-only way I found to work around the known limitation of
flex-boxes (nor grids) being able to shrink around wrapped contents.