Do not waste tokens on 'STOP'  (roll Next) - Fixes #22
Thanks to the Vercel team (@jridgewell), an interruption of the stream on the client
side will lead to the cancellation of the TransformStream on the servers side, which
in turns cancels the open fetch() to the upstream. This was a long needed change and
we are happy to report it works well.

Related: #114
 - https://github.com/vercel-labs/ai/issues/90
 - https://github.com/vercel/edge-runtime/pull/428
 - https://github.com/trpc/trpc/issues/4586 (enormous thanks to the tRPC team to issue
   a quick release as well)